Retain availability
Otherwise when sensors are first initialised to Home Assistant, HA
(correctly) waits for the next time the next availability state changes.

It works the second time kiot is launched, because in that case HA
already knows the entity has registered the availability topic so is
waiting for it.

If the state is maintained, HA uses the retained value.
The will message will take care of cleanup.